backend/requirements.txt
# FastAPI and Web Server
fastapi==0.115.4
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database - Supabase (includes PostgreSQL client)
supabase==2.9.1

# Redis - Upstash
redis==5.2.0
upstash-redis==0.15.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic==2.10.2
pydantic-settings==2.6.1

# Email
resend==2.19.0

# Utilities
python-dateutil==2.9.0
pytz==2024.2

# Reports
pandas==2.2.3
openpyxl==3.1.5

# WebSocket
python-socketio==5.11.4

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
